move api code to separate file
